Move `Occupancy` to separate file
Circular imports from work-stealing
